+++ test/foundry/AqueductV1Pair.reserves.t.sol
@@ -40,7 +40,7 @@ contract AqueductV1PairReservesTest is AqueductTester {
 
     function test_calculateFees_Basic() public {
         // Arrange
-        uint112 totalFlow = 10000;
+        uint96 totalFlow = 10000;
         uint32 timeElapsed = 1;
 
         // Act
@@ -52,7 +52,7 @@ contract AqueductV1PairReservesTest is AqueductTester {
 
     function test_calculateFees_NoFlow() public {
         // Arrange
-        uint112 totalFlow = 0;
+        uint96 totalFlow = 0;
         uint32 timeElapsed = 100;
 
         // Act
@@ -64,7 +64,7 @@ contract AqueductV1PairReservesTest is AqueductTester {
 
     function test_calculateFees_NoTimeElasped() public {
         // Arrange
-        uint112 totalFlow = 10000;
+        uint96 totalFlow = 10000;
         uint32 timeElapsed = 0;
 
         // Act
@@ -76,7 +76,7 @@ contract AqueductV1PairReservesTest is AqueductTester {
 
     function test_calculateFees_LongTime() public {
         // Arrange
-        uint112 totalFlow = 10000;
+        uint96 totalFlow = 10000;
         uint32 timeElapsed = 100;
 
         // Act
@@ -88,7 +88,7 @@ contract AqueductV1PairReservesTest is AqueductTester {
 
     function test_calculateFees_MaxValues() public {
         // Arrange
-        uint112 totalFlow = uint112(sf.cfa.MAXIMUM_FLOW_RATE()); // max 96 bits
+        uint96 totalFlow = uint96(sf.cfa.MAXIMUM_FLOW_RATE()); // max 96 bits
         uint32 timeElapsed = type(uint16).max; // max == 112 - 96 = 16 bits
 
         uint112 expectedFee = uint112((uint256(totalFlow) * timeElapsed * TWAP_FEE) / 10000);
@@ -105,7 +105,7 @@ contract AqueductV1PairReservesTest is AqueductTester {
         aqueductV1PairHarness.exposed_calculateFees(totalFlow, timeElapsed);
     }
 
-    function testFuzz_calculateFees(uint112 totalFlow, uint32 timeElapsed) public {
+    function testFuzz_calculateFees(uint96 totalFlow, uint32 timeElapsed) public {
         // Arrange
         vm.assume(totalFlow < sf.cfa.MAXIMUM_FLOW_RATE());
 
@@ -128,7 +128,7 @@ contract AqueductV1PairReservesTest is AqueductTester {
 
     function test_calculateReserveAmountSinceTime_Basic() public {
         // Arrange
-        uint112 totalFlow = 10000;
+        uint96 totalFlow = 10000;
         uint32 timeElapsed = 12;
         uint112 expectedReserve = (totalFlow * timeElapsed * (10000 - TWAP_FEE)) / 10000;
 
@@ -141,7 +141,7 @@ contract AqueductV1PairReservesTest is AqueductTester {
 
     function test_calculateReserveAmountSinceTime_NoFlow() public {
         // Arrange
-        uint112 totalFlow = 0;
+        uint96 totalFlow = 0;
         uint32 timeElapsed = 100;
 
         // Act
@@ -153,7 +153,7 @@ contract AqueductV1PairReservesTest is AqueductTester {
 
     function test_calculateReserveAmountSinceTime_NoTimeElapsed() public {
         // Arrange
-        uint112 totalFlow = 10000;
+        uint96 totalFlow = 10000;
         uint32 timeElapsed = 0;
 
         // Act
@@ -165,7 +165,7 @@ contract AqueductV1PairReservesTest is AqueductTester {
 
     function test_calculateReserveAmountSinceTime_MaxValues() public {
         // Arrange
-        uint112 totalFlow = uint112(sf.cfa.MAXIMUM_FLOW_RATE()); // max 96 bits
+        uint96 totalFlow = uint96(sf.cfa.MAXIMUM_FLOW_RATE()); // max 96 bits
         uint32 timeElapsed = type(uint16).max; // max == 112 - 96 = 16 bits
 
         uint112 expectedReserveAmount = uint112((uint256(totalFlow) * timeElapsed * (10000 - TWAP_FEE)) / 10000);
@@ -182,7 +182,7 @@ contract AqueductV1PairReservesTest is AqueductTester {
         aqueductV1PairHarness.exposed_calculateReserveAmountSinceTime(totalFlow, timeElapsed);
     }
 
-    function testFuzz_calculateReserveAmountSinceTime(uint112 totalFlow, uint32 timeElapsed) public {
+    function testFuzz_calculateReserveAmountSinceTime(uint96 totalFlow, uint32 timeElapsed) public {
         // Arrange
         vm.assume(totalFlow < (sf.cfa.MAXIMUM_FLOW_RATE()));
 
@@ -207,8 +207,8 @@ contract AqueductV1PairReservesTest is AqueductTester {
     function test_calculateReservesBothFlows_Basic() public {
         // Arrange
         uint256 kLast = 1 * 10 ** 36;
-        uint112 totalFlow0 = 1 * 10 ** 18;
-        uint112 totalFlow1 = 1 * 10 ** 18;
+        uint96 totalFlow0 = 1 * 10 ** 18;
+        uint96 totalFlow1 = 1 * 10 ** 18;
         uint32 timeElapsed = 12;
         uint112 reserve0 = 1 * 10 ** 18;
         uint112 reserve1 = 1 * 10 ** 18;
@@ -236,8 +236,8 @@ contract AqueductV1PairReservesTest is AqueductTester {
     function test_calculateReservesBothFlows_ZeroFlows() public {
         // Arrange
         uint256 _kLast = 1 * 10 ** 36;
-        uint112 totalFlow0 = 0;
-        uint112 totalFlow1 = 0;
+        uint96 totalFlow0 = 0;
+        uint96 totalFlow1 = 0;
         uint32 timeElapsed = 12;
         uint112 _reserve0 = 1 * 10 ** 18;
         uint112 _reserve1 = 1 * 10 ** 18;
@@ -260,8 +260,8 @@ contract AqueductV1PairReservesTest is AqueductTester {
     function test_calculateReservesBothFlows_NoTimeElapsed() public {
         // Arrange
         uint256 _kLast = 1 * 10 ** 36;
-        uint112 totalFlow0 = 1 * 10 ** 18;
-        uint112 totalFlow1 = 1 * 10 ** 18;
+        uint96 totalFlow0 = 1 * 10 ** 18;
+        uint96 totalFlow1 = 1 * 10 ** 18;
         uint32 timeElapsed = 0;
         uint112 _reserve0 = 1 * 10 ** 18;
         uint112 _reserve1 = 1 * 10 ** 18;
@@ -284,8 +284,8 @@ contract AqueductV1PairReservesTest is AqueductTester {
     function testFuzz_calculateReservesBothFlows(
         uint112 reserve0,
         uint112 reserve1,
-        uint112 totalFlow0,
-        uint112 totalFlow1,
+        uint96 totalFlow0,
+        uint96 totalFlow1,
         uint32 timeElapsed
     ) public {
         // Arrange
@@ -319,7 +319,7 @@ contract AqueductV1PairReservesTest is AqueductTester {
     function test_calculateReservesFlow0_Basic() public {
         // Arrange
         uint256 kLast = 1 * 10 ** 36;
-        uint112 totalFlow0 = 1 * 10 ** 18;
+        uint96 totalFlow0 = 1 * 10 ** 18;
         uint32 timeElapsed = 12;
         uint112 reserve0 = 1 * 10 ** 18;
 
@@ -345,7 +345,7 @@ contract AqueductV1PairReservesTest is AqueductTester {
     function test_calculateReservesFlow0_NoFlow() public {
         // Arrange
         uint256 kLast = 1 * 10 ** 36;
-        uint112 totalFlow0 = 0;
+        uint96 totalFlow0 = 0;
         uint32 timeElapsed = 12;
         uint112 reserve0 = 1 * 10 ** 18;
 
@@ -369,7 +369,7 @@ contract AqueductV1PairReservesTest is AqueductTester {
     function test_calculateReservesFlow0_NoElapsedTime() public {
         // Arrange
         uint256 kLast = 1 * 10 ** 36;
-        uint112 totalFlow0 = 1 * 10 ** 18;
+        uint96 totalFlow0 = 1 * 10 ** 18;
         uint32 timeElapsed = 0;
         uint112 reserve0 = 1 * 10 ** 18;
 
@@ -391,7 +391,7 @@ contract AqueductV1PairReservesTest is AqueductTester {
     }
 
     function testFuzz_calculateReservesFlow0(
-        uint112 totalFlow0,
+        uint96 totalFlow0,
         uint32 timeElapsed,
         uint112 reserve0,
         uint112 reserve1
@@ -417,7 +417,7 @@ contract AqueductV1PairReservesTest is AqueductTester {
     function test_calculateReservesFlow1_Basic() public {
         // Arrange
         uint256 kLast = 1 * 10 ** 36;
-        uint112 totalFlow1 = 1 * 10 ** 18;
+        uint96 totalFlow1 = 1 * 10 ** 18;
         uint32 timeElapsed = 12;
         uint112 reserve1 = 1 * 10 ** 18;
 
@@ -443,7 +443,7 @@ contract AqueductV1PairReservesTest is AqueductTester {
     function test_calculateReservesFlow1_NoFlow() public {
         // Arrange
         uint256 kLast = 1 * 10 ** 36;
-        uint112 totalFlow1 = 0;
+        uint96 totalFlow1 = 0;
         uint32 timeElapsed = 12;
         uint112 reserve1 = 1 * 10 ** 18;
 
@@ -467,7 +467,7 @@ contract AqueductV1PairReservesTest is AqueductTester {
     function test_calculateReservesFlow1_NoElapsedTime() public {
         // Arrange
         uint256 kLast = 1 * 10 ** 36;
-        uint112 totalFlow1 = 1 * 10 ** 18;
+        uint96 totalFlow1 = 1 * 10 ** 18;
         uint32 timeElapsed = 0;
         uint112 reserve1 = 1 * 10 ** 18;
 
@@ -489,7 +489,7 @@ contract AqueductV1PairReservesTest is AqueductTester {
     }
 
     function testFuzz_calculateReservesFlow1(
-        uint112 totalFlow1,
+        uint96 totalFlow1,
         uint32 timeElapsed,
         uint112 reserve0,
         uint112 reserve1
